+++ History.md
@@ -1,3 +1,8 @@
+0.0.3 - September 11, 2013
+--------------------------
+* added .value() getter
+* don't trigger change events on the same value
+
 0.0.2 - July 31, 2013
 ---------------------
 * added `name` method